backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0.1
anthropic==0.34.2
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
